backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2
django==4.1.1
django-adminlte-3==0.1.6
django-jazzmin==2.5.0
djangorestframework==3.13.1
pillow==9.2.0
pymysql==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
